requirements.txt
numpy
msgpack-rpc-python
numpy
airsim
controller
inputs
configargparse

# YOLOv5 requirements
# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
Pillow>=9.4.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.232

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
